chore(wallet): move getPriceIdForToken to pricing-utils
