+++ components/brave_wallet_ui/page/screens/composer_ui/to_asset/to_asset.tsx
@@ -8,8 +8,10 @@ import { skipToken } from '@reduxjs/toolkit/query/react'
 
 // Utils
 import { getLocale, splitStringForTag } from '../../../../../common/locale'
-import { computeFiatAmount } from '../../../../utils/pricing-utils'
-import { getPriceIdForToken } from '../../../../utils/api-utils'
+import {
+  computeFiatAmount,
+  getPriceIdForToken
+} from '../../../../utils/pricing-utils'
 import {
   getDominantColorFromImageURL //
 } from '../../../../utils/style.utils'
